+++ src/cancellation-token.js
@@ -2,87 +2,126 @@ import OperationCancellationError from './operation-cancellation-error.js';
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
-class CancellationToken
+function tryToMakeAbortController ()
 {
-	constructor (source = null)
+	return typeof AbortController === 'function' ? new AbortController() : null;
+}
+
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+export const Cancel = Symbol('Cancel');
+
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+export default class CancellationToken
+{
+	static None = new CancellationToken({
+		isCancelable : false
+	});
+
+	#reason                  = null;
+	#isCancelable            = false;
+	#isCancellationRequested = false;
+	#callbacks               = [];
+	#promise                 = null;
+	#aborter                 = tryToMakeAbortController();
+
+	constructor ({
+		isCancelable = true
+	} = {})
 	{
-		Object.defineProperty(this, 'source', {
-			value : source
-		});
+		this.#isCancelable = isCancelable;
 
-		Object.defineProperty(this, 'callbacks', {
-			value : []
-		});
-
-		Object.defineProperty(this, 'promise', {
-			value : new Promise((resolve, reject) =>
-			{
-				this.register(reject);
-			})
-		});
-
-		this.promise.catch(() => // eslint-disable-line promise/prefer-await-to-then
+		this.#promise = new Promise((resolve, reject) =>
 		{
-			// Here to prevent unhandled rejections.
+			this.register(reject);
+		});
+
+		this.#promise.catch(() =>
+		{
+			// Prevent unhandled rejections.
 		});
 	}
 
-	get canBeCancelled ()
+	get reason ()
 	{
-		return this.source !== null;
+		return this.#reason;
+	}
+
+	get isCancelable ()
+	{
+		return this.#isCancelable;
 	}
 
 	get isCancellationRequested ()
 	{
-		return this.canBeCancelled && this.source.isCancellationRequested;
+		return this.#isCancellationRequested;
 	}
 
 	register (callback)
 	{
-		if (this.canBeCancelled)
+		if (this.#isCancelable)
 		{
-			if (this.isCancellationRequested)
+			if (this.#isCancellationRequested)
 			{
-				callback();
+				callback(
+					new OperationCancellationError(this.#reason)
+				);
 			}
 			else
 			{
-				this.callbacks.push(callback);
+				this.#callbacks.push(callback);
 			}
 		}
+
+		return callback;
 	}
 
 	deregister (callback)
 	{
-		const i = this.callbacks.indexOf(callback);
+		const i = this.#callbacks.indexOf(callback);
 
 		if (i >= 0)
 		{
-			this.callbacks.splice(i, 1);
+			this.#callbacks.splice(i, 1);
 		}
 	}
 
 	throwIfCancellationRequested ()
 	{
-		if (this.isCancellationRequested)
+		if (this.#isCancellationRequested)
 		{
-			throw new OperationCancellationError();
+			throw new OperationCancellationError(this.#reason);
 		}
+	}
+
+	toAbortSignal ()
+	{
+		return this.#aborter?.signal;
 	}
 
 	then (onFulfilled, onRejected)
 	{
-		if (this.canBeCancelled)
+		if (this.#isCancelable)
 		{
-			return this.promise.then(onFulfilled, onRejected); // eslint-disable-line promise/prefer-await-to-then
+			return this.#promise.then(onFulfilled, onRejected);
 		}
 
 		return this;
 	}
+
+	[Cancel] (reason)
+	{
+		if (!this.#isCancellationRequested)
+		{
+			this.#isCancellationRequested = true;
+			this.#reason                  = reason;
+
+			this.#callbacks.splice(0).forEach(callback => callback(
+				new OperationCancellationError(this.#reason)
+			));
+
+			this.#aborter?.abort(reason);
+		}
+	}
 }
-
-CancellationToken.None = new CancellationToken();
-
-// - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
-export default CancellationToken;