+++ you-are-cancelled.d.ts
@@ -31,7 +31,7 @@ export class CancellationToken implements PromiseLike<void>
 	/**
 	 * Indicates whether the token is capable of being in a cancelled state.
 	 */
-	readonly canBeCancelled : boolean;
+	readonly isCancelable : boolean;
 
 	/**
 	 * Indicates whether a cancellation has been requested by its associative source.
@@ -39,18 +39,27 @@ export class CancellationToken implements PromiseLike<void>
 	readonly isCancellationRequested : boolean;
 
 	/**
+	 * Reflects the reason for why the cancellation was requested.
+	 *
+	 * If a cancellation has not been requested or a reason was not provided at cancellation then this will be `null`.
+	 */
+	readonly reason : string | null;
+
+	/**
 	 * Registers a callback function to execute when a cancellation has been requested.
 	 *
 	 * @param callback A function that you want to be executed when a cancellation has been requested.
+	 *
+	 * @returns The provided callback function.
 	 */
-	register(callback : OperationCancellationCallback) : void;
+	register(callback : OperationCancellationCallback) : OperationCancellationCallback;
 
 	/**
 	 * Deregisters a previously registered callback that would execute when a cancellation has been requested.
 	 *
 	 * @param callback The function that you no longer want to be executed when a cancellation has been requested.
 	 */
-	deregister(callback) : void;
+	deregister(callback : OperationCancellationCallback) : void;
 
 	/**
 	 * Throws an `OperationCancellationError` if a cancellation has already been requested.
@@ -67,9 +76,16 @@ export class CancellationToken implements PromiseLike<void>
 	throwIfCancellationRequested() : void;
 
 	/**
+	 * Converts this token into an `AbortSignal` object.
+	 *
+	 * @returns An `AbortSignal` object that reflects the cancellation state of this token.
+	 */
+	toAbortSignal() : AbortSignal;
+
+	/**
 	 * @inheritdoc
 	 */
-	then (onFulfilled ?: ((value : void) => any | PromiseLike<any>) | null | undefined, onRejected ?: ((reason : OperationCancellationError) => any | PromiseLike<any>) | null | undefined) : PromiseLike<any>;
+	then(onFulfilled ?: ((value : void) => any | PromiseLike<any>) | null | undefined, onRejected ?: ((reason : OperationCancellationError) => any | PromiseLike<any>) | null | undefined) : PromiseLike<any>;
 
 	/**
 	 * A dummy token that is not capable of being in a cancelled state.
@@ -101,5 +117,5 @@ export class CancellationTokenSource
 	 *
 	 * @param reason An optional explanation as to why the cancellation is being requested.
 	 */
-	cancel (message? : string) : void;
+	cancel(reason? : string) : void;
 }